Improve project root searching (#21)
- Prefer starting search in the dir containing the executable
- Fallback to previous method
- Changes the signature of findProjectRoot() (breaking change)

Previously we were using the #file directive to get a file path
inside the source tree of the project. This is resolved at compile-time
and therefore is incorrect if the project gets moved after compilation.

Since we have discovered a reliable way to determine the executable
location, it makes sense to use that as a starting point instead.